CUSBOSFW-65 Ability to disable ssl at runtime
We needed to disable ssl while trying to get the initial time.
If we didn't have the right time, ssl would fail with an expire
error, but we needed ssl to get the right time

To get around this, we disable ssl right before the time fetch
and re-enable it right after.